Prettied up newsfeed a bit. Filtering out certain posts that don't translate well
Posts that don't translate well are things like, someone commented on my
friend's relationship status from a month ago. Or users sharing a page,
no message, and not like a post with a link, like a facebook sharing of
a page.

In this commit I've whitelisted types of posts that seem to work well,
which is most, and also filter out posts without a message.